Revert "Delete 20210521043233_createTablesTable.js"
This reverts commit a68b8ba8e42bf907c4378918010dfc4ae42c631e.

Merge conflict with local repository.

The commit a68b8ba diverged from the local branch, but remotely.
I made 4 new local commits and tried to push to the remote origin.
Merge conflict: this commit does not exist in my local repo.
(I had to un-delete the file because it corrupted my migration directory
to have it deleted).
But since it was deleted remotely, the local branch diverged.
The last pushed commit on the remote needs to be removed
so the HEAD can be rebased to point to the last shared commit.
Then I can push the rest of my local commits (I hope).