Merge conflict between local and remote repo.
I deleted a migration locally and remotely.
I then added that migration back locally because
without it my migration directory was corrupt.
I did not add it back remotely, and now my local branch is
4 commits ahead of my origin/main branch.
This is an attempt to override the last commit in origin/main
branch and take it back to the last shared commit with my
current local branch, then push my local changes upstream again.